refactor(core): group text geometry helper
Move text-run geometry helpers into the existing text domain folder so the core root keeps fewer low-level text utilities.
