+++ src/core/text/geometry.ts
